Change data type of a2s array in comparative structure prediction
- This is a major change that increases the storage size of the
  alignment-column to sequence position mapping array 'a2s'. In early
  implementations, it was only used in RNAalifold and related functions,
  i.e. global structure prediction where alignment columns never
  exceeded about 30,000. Thus an 'unsigned short' storage type seemed
  adequat. However, Now that it is used in local folding variants of our
  implementations which do not have this low limit of alignment columns
  it easily exceeds storage size, thereby producing integer overflows.
  Consequently, this array is now of type 'unsigned int' which allows us
  to process alignments up to some hundred millions of columns in
  length.

+++ src/ViennaRNA/alphabet.h
@@ -91,7 +91,7 @@ void vrna_aln_encode( const char *sequence,
                       short **s5_p,
                       short **s3_p,
                       char **ss_p,
-                      unsigned short **as_p,
+                      unsigned int **as_p,
                       vrna_md_t *md);
 
 /**
